Updated routes sructure - there was starting to be too many in one file.
