+++ app/routes/__init__.py
@@ -1,0 +1,9 @@
+
+
+from app.routes.routes import bp
+from app.routes.auth_routes import auth_bp
+from app.routes.car_routes import car_bp
+from app.routes.info_routes import info_bp
+from app.routes.student_routes import student_bp
+from app.routes.teacher_routes import teacher_bp
+from app.routes.admin_routes import admin_bp
