+++ app/templates/add_module.html
@@ -17,7 +17,7 @@ content %}
     </button>
   </form>
 
-  <a href="{{ url_for('core.student_dashboard') }}" class="btn"
+  <a href="{{ url_for('student.student_dashboard') }}" class="btn"
     >Back to Dashboard</a
   >
   {% endblock %}
