+++ app/templates/base.html
@@ -35,7 +35,7 @@
             </li>
             {% elif current_user.role == 'student' %}
             <li>
-              <a href="{{ url_for('core.student_dashboard') }}"
+              <a href="{{ url_for('student.student_dashboard') }}"
                 >Student Dashboard</a
               >
             </li>
