+++ app/templates/upload_profile_picture.html
@@ -38,7 +38,7 @@
 
   <!-- Back to Dashboard -->
   <a
-    href="{{ url_for('core.' + current_user.role + '_dashboard') }}"
+    href="{{ dashboard_url }}"
     class="btn btn-success back-to-dashboard"
   >
     Back to Dashboard
